[BT-52] grid example styles. Moved grouped component files into dedicated folder in examples
