+++ src/examples/_partials/_nav.jade
@@ -21,7 +21,7 @@ script.
       var value = source[i];
       var markup = [
         '<a class="examplesNavButton" href="',
-        '/' + rootDir + '/' + value + '.html"',
+        '/' + rootDir + '/' + value + '/' + value + '.html"',
         '>' + value + '</a>'
       ].join('');
       var button = $(markup);
