+++ src/examples/components/Grid/Grid.jade
@@ -1,5 +1,5 @@
 
-extends ../_partials/_componentLayout.jade
+extends ../../_partials/_componentLayout
 
 block title
   title Grid
