+++ src/examples/components/Grid/_grid.scss
@@ -1,0 +1,11 @@
+
+.gridExample__container {
+  box-shadow: none;
+  border: 1px solid $componentStrokeColor;
+}
+
+.gridExample__headerCell,
+.gridExample__bodyCell,
+.gridExample__footerCell {
+  padding: 10px;
+}
