--- src/examples/components/grid.jade
@@ -1,18 +1,0 @@
-
-extends ../_partials/_componentLayout.jade
-
-block title
-  title Grid
-
-block content
-
-  .exampleContainer.exampleContainer--clear
-    #grid
-
-block component
-  div.
-    <script type="text/jsx">
-      window.renderComponent('grid', (
-        <GridExample />
-      ));
-    </script>
