+++ src/examples/index.scss
@@ -93,3 +93,5 @@
 .exampleContainer--clear {
   background: transparent;
 }
+
+@import 'components/Grid/grid';
