+++ src/examples/integrations/tableView/tableView.jade
@@ -1,5 +1,5 @@
 
-extends ../_partials/_integrationLayout.jade
+extends ../../_partials/_integrationLayout
 
 block title
   title Table View
